[sql][Create_Table_Marks.sql]

+++ Deanary/Create_Table_Marks.sql
@@ -5,7 +5,7 @@ GO
 CREATE TABLE MARKS(
 	id int NOT NULL IDENTITY (1, 1)	PRIMARY KEY,
 	grade nvarchar(16),
-	min_mark_number int NOT NULL,
-	max_mark_number int NOT NULL,
+	min_mark_number tinyint NOT NULL,
+	max_mark_number tinyint NOT NULL,
 	differentiated bit
 );
